Optionally configure Telegraf on remote PuppetDB nodes
Prior to this commit, there was no way to configure the influxdb node
when running telegraf on a remote PuppetDB node. It would default to
`localhost` and one would have to add the class external to this module.
This commit adds a new `influxdb_host` parameter that configures
telegraf to send the data to that node.